# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Monee Local Ledger API
  version: 1.2.0
  description: Unified product users with explicitly linked Google/GitHub identities.
    System admission is separate from ledger ownership/editor/viewer memberships.
    All data access is enforced by Go. Feishu is reserved for a future provider adapter.
    Local loopback HTTP only; cloud deployment is not yet supported. Cookie-authenticated
    writes require X-Monee-CSRF. Application database schema version 2.
servers:
- url: http://127.0.0.1:4173
security:
- localBearer: []
- browserSession: []
paths:
...
  "/api/v1/dashboard":
    get:
      operationId: dashboard
      description: Statistics cover the entire month. Search and page only filter
        the transaction list.
      parameters:
      - in: query
        name: month
        description: YYYY-MM between 1900 and 2200; empty or omitted selects this
          month in Asia/Shanghai.
        schema:
          type: string
      - in: query
        name: q
        description: Case-insensitive literal substring of merchant, category, source,
          account or note.
        schema:
          type: string
          maxLength: 100
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          maximum: 100000
          default: 1
      - in: header
        name: X-Monee-Ledger
        required: false
        description: Explicit target ledger ID. Clients always send this. Omission
          is accepted only when the user belongs to exactly one ledger; otherwise
          ledger_required. Membership is checked for each operation.
        schema:
          type: string
      responses:
        '200':
          description: Monthly aggregates and at most 50 transactions, sorted by date
            and creation time descending, then ID.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Dashboard"
        default:
          "$ref": "#/components/responses/Error"
  "/api/v1/transactions":
    post:
      operationId: createTransaction
      description: Same key and normalized input return the same transaction, including
        after a lost response. Changing the input under that key returns 409.
      parameters:
      - in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
      - in: header
        name: X-Monee-CSRF
        required: false
        description: Required when authenticating this write with a browser cookie;
          use the proof from auth/me. Not required for native Bearer sessions.
        schema:
          type: string
      - in: header
        name: X-Monee-Ledger
        required: false
        description: Explicit target ledger ID. Clients always send this. Omission
          is accepted only when the user belongs to exactly one ledger; otherwise
          ledger_required. Membership is checked for each operation.
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransactionInput"
      responses:
        '201':
          description: Persisted transaction (also returned for an idempotent replay).
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Transaction"
        default:
          "$ref": "#/components/responses/Error"
  "/api/v1/imports/preview":
    post:
      operationId: previewImport
      description: |
        Standard UTF-8 CSV, optional BOM, at most 2 MiB and 1000 records.
        Required columns are date,type,amount,currency,merchant,source.
        Optional columns are category,account,external_id,note.
        Structural or row errors are returned in errors with HTTP 200; do not commit such a result.
        Batches without validation/identity conflicts and their raw CSV are persisted.
        Transactions are not created by preview.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - csv
              properties:
                filename:
                  type: string
                  description: Basename is retained; at most 200 characters after
                    basename extraction.
                csv:
                  type: string
      responses:
        '200':
          description: Preview with errors, duplicate/similar status or alreadyCommitted
            indication.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ImportPreview"
        default:
          "$ref": "#/components/responses/Error"
      parameters:
      - in: header
        name: X-Monee-CSRF
        required: false
        description: Required when authenticating this write with a browser cookie;
          use the proof from auth/me. Not required for native Bearer sessions.
        schema:
          type: string
      - in: header
        name: X-Monee-Ledger
        required: false
        description: Explicit target ledger ID. Clients always send this. Omission
          is accepted only when the user belongs to exactly one ledger; otherwise
          ledger_required. Membership is checked for each operation.
        schema:
          type: string
  "/api/v1/imports/{id}/commit":
    post:
      operationId: commitImport
      description: |
        Commit the complete preview atomically. Ledger changes since preview cause 409.
        Similar records require confirmSimilar=true, acknowledging they are independent transactions.
        This does not merge cross-source records. Already committed batches return the original result.
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: header
        name: X-Monee-CSRF
        required: false
        description: Required when authenticating this write with a browser cookie;
          use the proof from auth/me. Not required for native Bearer sessions.
        schema:
          type: string
      - in: header
        name: X-Monee-Ledger
        required: false
        description: Explicit target ledger ID. Clients always send this. Omission
          is accepted only when the user belongs to exactly one ledger; otherwise
          ledger_required. Membership is checked for each operation.
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                ledgerVersion:
                  type: integer
                  format: int64
                  default: 0
                confirmSimilar:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Original commit result; added/skipped are counts for that batch,
            not this HTTP request.
          content:
            application/json:
              schema:
                type: object
                required:
                - importId
                - added
                - skipped
                - month
                properties:
                  importId:
                    type: string
                    format: uuid
                  added:
                    type: integer
                    minimum: 0
                  skipped:
                    type: integer
                    minimum: 0
                  month:
                    type: string
                    description: Latest business month in this batch.
        default:
          "$ref": "#/components/responses/Error"
  "/api/v1/template":
    get:
      operationId: csvTemplate
      parameters:
      - in: query
        name: sample
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Standard CSV header, with three synthetic records when sample=true.
          content:
            text/csv:
              schema:
                type: string
        default:
          "$ref": "#/components/responses/Error"
...
  "/auth/begin":
    get:
      operationId: beginBrowserLogin
      security: []
      parameters:
      - in: query
        name: ticket
        required: true
        schema:
          type: string
      responses:
        '303':
          description: Browser redirect. Invalid, expired or cancelled attempts redirect
            to a fixed result page, never an arbitrary return URL.
          headers:
            Location:
              schema:
                type: string
...
components:
  securitySchemes:
    localBearer:
      type: http
      scheme: bearer
      description: Opaque 12-hour application session obtained only by the native
        proof-bound OAuth exchange. The connection file contains no token.
    browserSession:
      type: apiKey
      in: cookie
      name: monee_session_eLaGr4oiqzKw
      description: HttpOnly SameSite=Lax session. The cookie name uses an origin hash;
        this name is for the default origin. Follow Set-Cookie when using a different
        port.
  responses:
    Error:
      description: 400 malformed JSON; 401 absent/expired session or invalid native
        proof; 403 data access denied, admin role required, CSRF or origin rejected;
        404 unknown resource; 409 identity/version conflict; 415 non-JSON write; 422
        validation; 429 too many login attempts; 503 provider unconfigured; 500 internal
        failure. Unknown routes may use plain-text 404/405.
      content:
        application/json:
          schema:
            type: object
            required:
            - code
            - message
            properties:
              code:
                type: string
              message:
                type: string
  schemas:
    Minor:
      type: string
      pattern: "^-?[0-9]+$"
      description: Signed 64-bit integer fen as a decimal string. Never a JSON number.
    TransactionInput:
      type: object
      additionalProperties: false
      required:
      - date
      - type
      - amount
      - currency
      - merchant
      - source
      properties:
        date:
          type: string
          format: date
          description: Valid Gregorian date between 1900 and 2200.
        type:
          type: string
          enum:
          - income
          - expense
          - 收入
          - 支出
        amount:
          type: string
          pattern: "^[0-9]{1,11}(\\.[0-9]{1,2})?$"
          description: Positive yuan, up to 10000000000.00; trimmed before validation.
        currency:
          type: string
          const: CNY
        merchant:
          type: string
          minLength: 1
          maxLength: 200
        category:
          type: string
          maxLength: 200
          description: Empty becomes 未分类.
        source:
          type: string
          minLength: 1
          maxLength: 200
        account:
          type: string
          maxLength: 200
          description: Empty becomes 待核实资金账户.
        externalId:
          type: string
          maxLength: 200
        note:
          type: string
          maxLength: 1000
    Transaction:
      type: object
      required:
      - id
      - date
      - type
      - amountMinor
      - currency
      - merchant
      - category
      - source
      - account
      - externalId
      - note
      - version
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        type:
          type: string
          enum:
          - income
          - expense
        amountMinor:
          "$ref": "#/components/schemas/Minor"
        currency:
          type: string
          const: CNY
        merchant:
          type: string
        category:
          type: string
        source:
          type: string
        account:
          type: string
          description: Currently a clearing account; does not imply a verified asset
            balance.
        externalId:
          type: string
        note:
          type: string
        version:
          type: integer
          format: int64
          minimum: 1
    Dashboard:
      type: object
      required:
      - ledgerId
      - version
      - month
      - today
      - months
      - incomeMinor
      - expenseMinor
      - sourceCount
      - totalCount
      - filteredCount
      - page
      - pageSize
      - categories
      - transactions
      properties:
        ledgerId:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
          minimum: 0
        month:
          type: string
        today:
          type: string
          format: date
        months:
          type: array
          items:
            type: string
          description: Descending; includes selected month and current month.
        incomeMinor:
          "$ref": "#/components/schemas/Minor"
          description: Nonnegative monthly income in fen.
        expenseMinor:
          "$ref": "#/components/schemas/Minor"
          description: Nonnegative monthly expense in fen.
        sourceCount:
          type: integer
          minimum: 0
        totalCount:
          type: integer
          minimum: 0
        filteredCount:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          const: 50
        categories:
          type: array
          description: Expense categories sorted by amount descending then name.
          items:
            type: object
            required:
            - name
            - amountMinor
            properties:
              name:
                type: string
              amountMinor:
                "$ref": "#/components/schemas/Minor"
                description: Positive expense total in fen.
        transactions:
          type: array
          maxItems: 50
          items:
            "$ref": "#/components/schemas/Transaction"
    ImportPreview:
      type: object
      required:
      - id
      - filename
      - ledgerVersion
      - rows
      - newCount
      - duplicateCount
      - similarCount
      - errors
      - alreadyCommitted
      properties:
        id:
          type: string
          description: Stable batch UUID; empty on parse failure.
        filename:
          type: string
        ledgerVersion:
          type: integer
          format: int64
          minimum: 0
        newCount:
          type: integer
          minimum: 0
          description: Includes similar records that require explicit confirmation.
        duplicateCount:
          type: integer
          minimum: 0
        similarCount:
          type: integer
          minimum: 0
        errors:
          type: array
          items:
            type: string
        alreadyCommitted:
          type: boolean
        rows:
          type: array
          maxItems: 1000
          items:
            type: object
            required:
            - line
            - record
            - status
            - message
            properties:
              line:
                type: integer
                minimum: 2
                description: Physical CSV line where the record begins.
              record:
                "$ref": "#/components/schemas/Transaction"
              status:
                type: string
                enum:
                - new
                - duplicate
                - similar
                - conflict
              message:
                type: string
...